chore(bower): add all dist files to bower main
according to http://bower.io/docs/creating-packages/#bowerjson
the "main" key in a bower.json should include
"The primary acting files necessary to use your package"
as pointed out by @brad-sobie the webfontloader is currently
a hard dependency.
Also the css is required in order to have this project be usable.

ref #42